LayoutTests/editing/selection/ios/selection-clip-rect-in-overflow-scroller-expected.txt

Verifies that the selection highlight is clipped to overflow scrolling container. Select the first word in the container below and scroll down

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Here’s to the crazy ones. The misfits. The rebels. The troublemakers. The round pegs in the square holes. The ones who see things differently. They’re not fond of rules. And they have no respect for the status quo.

You can quote them, disagree with them, glorify or vilify them. About the only thing you can’t do is ignore them.

Because they change things. They push the human race forward. And while some may see them as the crazy ones, we see genius. Because the people who are crazy enough to think they can change the world, are the ones who do.

PASS getSelection().toString() became 'Here’s'
PASS selectionRects[0]?.width || 0 is 0
PASS selectionRects[0]?.height || 0 is 0
PASS successfullyParsed is true

TEST COMPLETE
